requirements.txt
streamlit
pillow
moviepy
python-docx
tqdm